+++ R/arg_character.R
@@ -1,7 +1,7 @@
 #' Check Character Argument
 #'
 #' @description
-#' Checks whether an argument is a character vector (`arg_character()`), a character scalar (`arg_string()`), or a factor (`arg_factor()`).
+#' Checks whether an argument is a character vector (`arg_character()`), a character scalar (`arg_string()`), a factor (`arg_factor()`), or an ordered factor (`arg_ordered()`).
 #'
 #' @inheritParams arg_is
 #'
@@ -10,7 +10,7 @@
 #'
 #' @inherit arg_is return
 #'
-#' @seealso [is.character()], [is.factor()], [rlang::is_string()]
+#' @seealso [is.character()], [is.factor()], [is.ordered()], [rlang::is_string()]
 #'
 #' @examples
 #' f <- function(z) {
@@ -27,7 +27,7 @@ arg_character <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
   if (!is.character(x)) {
 
     if (is_not_null(.msg)) {
-      err(.msg_eval(.msg), .call = .call)
+      err(.msg_eval(.msg), .call = .call, .envir = rlang::caller_env())
     }
 
     err("{.arg {(.arg)}} must be a character vector",
@@ -42,7 +42,12 @@ arg_string <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
   if (!is_scalar(x) || !is.character(x) || anyNA(x)) {
 
     if (is_not_null(.msg)) {
-      err(.msg_eval(.msg), .call = .call)
+      err(.msg_eval(.msg), .call = .call, .envir = rlang::caller_env())
+    }
+
+    if (is_scalar(x) && is.character(x)) {
+      err("{.arg {(.arg)}} must not be {.val {NA}}",
+          .call = .call)
     }
 
     err("{.arg {(.arg)}} must be a string",
@@ -57,10 +62,25 @@ arg_factor <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
   if (!is.factor(x)) {
 
     if (is_not_null(.msg)) {
-      err(.msg_eval(.msg), .call = .call)
+      err(.msg_eval(.msg), .call = .call, .envir = rlang::caller_env())
     }
 
     err("{.arg {(.arg)}} must be a factor",
+        .call = .call)
+  }
+}
+
+#' @export
+#' @rdname arg_character
+arg_ordered <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
+                       .call) {
+  if (!is.ordered(x)) {
+
+    if (is_not_null(.msg)) {
+      err(.msg_eval(.msg), .call = .call, .envir = rlang::caller_env())
+    }
+
+    err("{.arg {(.arg)}} must be an ordered factor",
         .call = .call)
   }
 }