+++ R/arg_dots_supplied.R
@@ -6,7 +6,7 @@
 #' @param \dots the `...` argument passed from a calling function. The argument is not evaluated. See Examples.
 #'
 #' @details
-#' `arg_dots_supplied()` checks whether arguments were passed to the `...` (i.e., "dots") argument of a function. These arguments are not evaluated. `arg_dots_supplied()` throws an error if [...length()] is 0. It can be useful for when a function must have additional arguments. For example, [arg_or()] and [when_supplied()] use [arg_dots_supplied()].
+#' `arg_dots_supplied()` checks whether arguments were passed to the `...` (i.e., "dots") argument of a function. These arguments are not evaluated. `arg_dots_supplied()` throws an error if [...length()] is 0. It can be useful for when a function must have additional arguments. For example, [when_not_null()] and [when_supplied()] use [arg_dots_supplied()].
 #'
 #' `arg_dots_not_supplied()` checks whether no arguments were passed to `...`, again without evaluating the arguments and using `...length()`. It can be useful when a function appears to allow further arguments (e.g., because it is a method of a generic), but when the author does not want the user to supply them.
 #'
@@ -27,7 +27,7 @@ arg_dots_supplied <- function(..., .msg = NULL, .call) {
   if (...length() == 0L) {
 
     if (is_not_null(.msg)) {
-      err(.msg_eval(.msg), .call = .call)
+      err(.msg_eval(.msg), .call = .call, .envir = rlang::caller_env())
     }
 
     err("an argument must be supplied to {.arg ...}",
@@ -41,7 +41,7 @@ arg_dots_not_supplied <- function(..., .msg = NULL, .call) {
   if (...length() != 0L) {
 
     if (is_not_null(.msg)) {
-      err(.msg_eval(.msg), .call = .call)
+      err(.msg_eval(.msg), .call = .call, .envir = rlang::caller_env())
     }
 
     err("no arguments may be supplied to {.arg ...}",