+++ R/arg_named.R
@@ -46,10 +46,20 @@ arg_named <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
       !all(nzchar(names(x)))) {
 
     if (is_not_null(.msg)) {
-      err(.msg_eval(.msg), .call = .call)
+      err(.msg_eval(.msg), .call = .call, .envir = rlang::caller_env())
     }
 
-    err("{.arg {(.arg)}} must have non-empty and non-{.val {NA}} names",
+    if (is_null(names(x)) || !all(nzchar(names(x)))) {
+      if (anyNA(names(x))) {
+        err("{.arg {(.arg)}} must not have empty names or {.val {NA}} values as names",
+            .call = .call)
+      }
+
+      err("{.arg {(.arg)}} must not have empty names",
+          .call = .call)
+    }
+
+    err("{.arg {(.arg)}} must not have {.val {NA}} values as names",
         .call = .call)
   }
 }
@@ -62,10 +72,20 @@ arg_colnamed <- function(x, .arg = rlang::caller_arg(x), .msg = NULL,
       !all(nzchar(colnames(x)))) {
 
     if (is_not_null(.msg)) {
-      err(.msg_eval(.msg), .call = .call)
+      err(.msg_eval(.msg), .call = .call, .envir = rlang::caller_env())
     }
 
-    err("{.arg {(.arg)}} must have non-empty and non-{.val {NA}} names",
+    if (is_null(colnames(x)) || !all(nzchar(colnames(x)))) {
+      if (anyNA(colnames(x))) {
+        err("{.arg {(.arg)}} must not have empty column names or {.val {NA}} values as column names",
+            .call = .call)
+      }
+
+      err("{.arg {(.arg)}} must not have empty column names",
+          .call = .call)
+    }
+
+    err("{.arg {(.arg)}} must not have {.val {NA}} values as column names",
         .call = .call)
   }
 }