+++ R/arg_unique.R
@@ -1,10 +1,10 @@
 #' Check Unique Argument
 #'
 #' @description
-#' Checks whether an argument contains only unique values.
+#' Checks whether an argument contains only unique (i.e., distinct) values.
 #'
 #' @inheritParams arg_is
-#' @param \dots further argument passed to [anyDuplicated()], which performs the checking.
+#' @param \dots further arguments passed to [anyDuplicated()], which performs the checking.
 #'
 #' @inherit arg_is return
 #'
@@ -25,7 +25,7 @@ arg_unique <- function(x, ..., .arg = rlang::caller_arg(x), .msg = NULL,
   if (is_not_null(x) && anyDuplicated(x, ...) != 0) {
 
     if (is_not_null(.msg)) {
-      err(.msg_eval(.msg), .call = .call)
+      err(.msg_eval(.msg), .call = .call, .envir = rlang::caller_env())
     }
 
     err("{.arg {(.arg)}} must only contain unique values",